refactor(modeling - order ) : 주문 공통 부분 패키지 이동 및 내용 수정
