--- src/main/java/kitchenpos/order/domain/JpaOrderRepository.java
@@ -1,8 +1,0 @@
-package kitchenpos.order.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
-}
