--- src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,16 +1,0 @@
-package kitchenpos.order.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface OrderRepository {
-    Order save(Order order);
-
-    Optional<Order> findById(UUID id);
-
-    List<Order> findAll();
-
-    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
-}
-
