+++ src/main/java/kitchenpos/orders/domain/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.orders.domain;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
