+++ src/test/java/kitchenpos/application/InMemoryOrderRepository.java
@@ -1,9 +1,7 @@
 package kitchenpos.application;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+
+import kitchenpos.eatinorder.domain.OrderTable;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -12,7 +10,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
-public class InMemoryOrderRepository implements OrderRepository {
+public class InMemoryOrderRepository  {
+    /*
     private final Map<UUID, Order> orders = new HashMap<>();
 
     @Override
@@ -36,5 +35,5 @@ public class InMemoryOrderRepository implements OrderRepository {
         return orders.values()
             .stream()
             .anyMatch(order -> order.getOrderTable().equals(orderTable) && order.getStatus() != status);
-    }
+    }*/
 }
